+++ attaques/Elfee/Poudre soporifique.cs
@@ -9,7 +9,7 @@ public class PoudreSoporifique : Attaque
         cout = 7;
         porteeMin = 1;
         porteeMax = 1;
-        typeCible = (int)Jeu.CibleType.persoEnnemy;
+        typeCible = Jeu.CibleType.persoEnnemy;
     }
 
     // Méthodes public
