+++ attaques/Elfee/Poudre stimulante.cs
@@ -11,7 +11,7 @@ public class PoudreStimulante : Attaque
         porteeMin = 1;
         porteeMax = 3;
         ligneDeVue = true;
-        typeCible = (int)Jeu.CibleType.persoFriendly;
+        typeCible = Jeu.CibleType.persoFriendly;
     }
 
     // Méthodes public
