+++ attaques/Fantomage/Buff.cs
@@ -11,7 +11,7 @@ public class Buff : Attaque
         porteeMin = 1;
         porteeMax = 3;
         ligneDeVue = false;
-        typeCible = (int)Jeu.CibleType.persoFriendly;
+        typeCible = Jeu.CibleType.persoFriendly;
     }
 
     // Méthodes public
