+++ attaques/Fantomage/Case de rapatriement.cs
@@ -11,7 +11,7 @@ public class CaseDeRapatriement : Attaque
         porteeMin = 1;
         porteeMax = 3;
         ligneDeVue = false;
-        typeCible = (int)Jeu.CibleType.caseDeRapatriement;
+        typeCible = Jeu.CibleType.caseDeRapatriement;
         myCase = null;
     }
