+++ attaques/Fantomage/Rappel.cs
@@ -10,7 +10,7 @@ public class Rappel : Attaque
         porteeMin = 0;
         porteeMax = 5;
         ligneDeVue = false;
-        typeCible = (int)Jeu.CibleType.rappel;
+        typeCible = Jeu.CibleType.rappel;
     }
 
     // Méthodes public
