+++ attaques/Roninja/Dérobade.cs
@@ -9,7 +9,7 @@ public class Derobade : Attaque
         cout = 4;
         porteeMin = 1;
         porteeMax = 1;
-        typeCible = (int)Jeu.CibleType.derobade;
+        typeCible = Jeu.CibleType.derobade;
     }
 
     // Méthodes public
