+++ attaques/Roninja/Memoire.cs
@@ -9,7 +9,7 @@ public class Memoire : Attaque
         cout = 1;
         porteeMin = 0;
         porteeMax = 0;
-        typeCible = (int)Jeu.CibleType.memoire;
+        typeCible = Jeu.CibleType.memoire;
         tp = null;
     }
